+++ src/app/(app)/import/page.tsx
@@ -4,7 +4,7 @@ import { ImportView } from "@/components/import/import-view";
 export default async function ImportPage() {
   return (
     <div>
-      <Topbar title="Import transactions" />
+      <Topbar title="Import Transactions" />
       <div className="p-4 lg:p-6 space-y-4">
         <ImportView />
       </div>